Restructure component directory hierarchy and enable routes
