+++ src/app/app.component.html
@@ -1,7 +1,1 @@
-<mat-toolbar color="primary">
-  Sign up Flow
-  <button mat-button>
-    <mat-icon>login</mat-icon>
-    Login
-  </button>
-</mat-toolbar>
+<router-outlet></router-outlet>
