--- src/app/component/landing/landing.component.html
@@ -1,1 +1,0 @@
-<p>landing works!</p>
