+++ src/app/routes/landing/landing.component.html
@@ -1,0 +1,8 @@
+<mat-toolbar color="primary">
+  Sign up Flow
+  <button mat-button>
+    <mat-icon>login</mat-icon>
+    Login
+  </button>
+</mat-toolbar>
+<p>Landing works!</p>
